feat(inventory): costo por bulto, margen, precio sugerido y stock decimal
- costPerBulk + unitsPerBulk + marginPct en productos
- costo unitario y precio sugerido calculados
- stock decimal (kg) en productos y movimientos
- carga masiva en dos pasos: productos (maestros) y precios
- precios identifican el producto por nombre/SKU/codigo de barras
- SKU autogenerado cuando viene vacio
- elimina router tRPC muerto de inventario

+++ apps/web/app/api/inventory/stock/route.ts
@@ -6,7 +6,7 @@ import { adjustStock, listMovements, listProducts } from "@apolo/module-inventor
 const schema = z.object({
   productId: z.string().uuid(),
   type: z.enum(["in", "out", "adjustment"]),
-  quantity: z.number().int().positive(),
+  quantity: z.number().positive(),
   notes: z.string().max(300).nullable().optional(),
 });
